Updated names to follow Object Orient syntax
